rejig and updates after review
